+++ sentinel/src/main/java/gov/va/health/api/sentinel/crawler/Crawler.java
@@ -171,7 +171,6 @@ public class Crawler {
       }
 
       addLinksFromBundle(payload);
-
     } else {
       resultBuilder.outcome(Outcome.INVALID_STATUS);
     }
